TIP-1232 Fix comments
